Change module to be downloaded into pyFAT
